refactor: keep everything in one folder
