+++ .gitignore
@@ -41,3 +41,6 @@ terraform.rc
 
 # Ignore environments 
 .venv/
+
+# Ignore cached files
+*__pycache__
