--- pyproject.toml
@@ -1,6 +1,0 @@
-[project]
-name = "takehome"
-version = "0.0.1"
-
-[tool.setuptools.packages.find]
-where = ["takehome"]
