+++ takehome/exercise1_losses_calculator.py
