+++ takehome/exercise3.md
