+++ takehome/generate_data.py
